Fix potential explicit cuda device errors

+++ scripts/ddpm.py
@@ -680,12 +680,12 @@ class UNet(DDPM):
             self.alphas_cumprod.shape[0] == self.num_timesteps
         ), "alphas have to be defined for each timestep"
 
-        to_torch = lambda x: x.to(self.cdevice)
+        to_torch = lambda x: x.to(self.cdevice, dtype=torch.float32)
         self.register_buffer1("betas", to_torch(self.betas))
         self.register_buffer1("alphas_cumprod", to_torch(self.alphas_cumprod))
         # ddim sampling parameters
         ddim_sigmas, ddim_alphas, ddim_alphas_prev = make_ddim_sampling_parameters(
-            alphacums=self.alphas_cumprod.cpu(),
+            alphacums=self.alphas_cumprod.to(torch.float32).cpu(),
             ddim_timesteps=self.ddim_timesteps,
             eta=ddim_eta,
             verbose=verbose,
